% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonlightR.R
\docType{data}
\name{dataFilt}
\alias{dataFilt}
\title{Gene Expression (Rnaseqv2) data from TCGA LUAD}
\format{A 13742x20 matrix}
\usage{
data(dataFilt)
}
\value{
a 13742x20 matrix
}
\description{
A data set containing the following data:
}
\details{
\itemize{
	\item dataFilt matrix with 13742 rows (genes) and 20 columns samples with TCGA's barcodes (10TP, 10NT)
}
}
\keyword{datasets}
