% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonlightR.R
\docType{data}
\name{dataGRN}
\alias{dataGRN}
\title{GRN gene regulatory network output}
\format{A large list of 2 elements}
\usage{
data(dataGRN)
}
\value{
a large list of 2 elements
}
\description{
output from GRN function
}
\details{
\itemize{
	\item dataGRN list of 2 elements miTFGenes, maxmi from GRN function
}
}
\keyword{datasets}
