% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonlightR.R
\docType{data}
\name{dataURA}
\alias{dataURA}
\title{Output example from function Upstram Regulator Analysis}
\format{
A data frame with 100 rows and 2 variables
}
\usage{
data(dataURA)
}
\value{
a 100x2 matrix
}
\description{
A data set containing the following data:
}
\details{
\itemize{
	\item dataURA matrix with 100 rows (genes) and 2 columns   
	 "apoptosis"              "proliferation of cells"
}
}
\keyword{datasets}
