% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonlightR.R
\docType{data}
\name{DEGsmatrix}
\alias{DEGsmatrix}
\title{DEG Differentially expressed genes}
\format{A 3502x5 matrix}
\usage{
data(DEGsmatrix)
}
\value{
the 3502x5 matrix
}
\description{
A data set containing the following data:
}
\details{
\itemize{
	\item DEGsmatrix matrix with 3502 rows (genes) and five columns  "logFC"  "logCPM" "LR"     "PValue" "FDR" 
}
}
\keyword{datasets}
