% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonlightR.R
\docType{data}
\name{DiseaseList}
\alias{DiseaseList}
\title{Information on 101 biological processes}
\format{
A list of 101 matrices
}
\usage{
data(DiseaseList)
}
\value{
list of 101 matrices
}
\description{
A data set containing the following data:
}
\details{
\itemize{
	\item DiseaseList list for 101 biological processes, each containing a matrix with five columns: ID, Genes.in.dataset, Prediction based on expression direction, Log ratio, Findings
}
}
\keyword{datasets}
