% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonlightR.R
\docType{data}
\name{EAGenes}
\alias{EAGenes}
\title{Information about genes}
\format{
A 20038x5 matrix
}
\usage{
data(EAGenes)
}
\value{
a 20038x5 matrix
}
\description{
A data set containing the following data:
}
\details{
\itemize{
	\item EAGenes matrix with 20038 rows (genes) and five columns   
	"ID"          "Gene"        "Description" "Location"    "Family"     
}
}
\keyword{datasets}
