% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonlightR.R
\docType{data}
\name{GDCprojects}
\alias{GDCprojects}
\title{Information on GDC projects}
\format{A character vector of 39 elements}
\usage{
data(GDCprojects)
}
\value{
character vector of 39 elements
}
\description{
A character vector of GDC projects:
}
\details{
\itemize{
	\item character vector for GDC projects.
}
}
\keyword{datasets}

