% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonlightR.R
\docType{data}
\name{geneInfo}
\alias{geneInfo}
\title{Information about genes for normalization}
\format{A data frame with 20531 rows and 3 variables}
\usage{
data(geneInfo)
}
\value{
a 20531x3 matrix
}
\description{
A data set containing the following data:
}
\details{
\itemize{
	\item geneInfo matrix with 20531 rows (genes) and 3 columns   
	"geneLength" "gcContent"  "chr"         
}
}
\keyword{datasets}

