% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonlightR.R
\docType{data}
\name{GEO_TCGAtab}
\alias{GEO_TCGAtab}
\title{Information on GEO data (and overlap with TCGA)#'
A data set containing the following data:}
\format{A 101x3 matrix}
\usage{
data(GEO_TCGAtab)
}
\value{
a 101x3 matrix
}
\description{
\itemize{
\item GEO_TCGAtab a 18x12 matrix that provides the GEO data set we matched to one of the 18 given TCGA cancer types 
}
}
\keyword{datasets}

