% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.data.GEO.R
\name{get.data.GEO}
\alias{get.data.GEO}
\title{get.data.GEO}
\usage{
get.data.GEO(GEOobject = "GSE39004", platform = "GPL6244",
  TCGAtumor = NULL)
}
\arguments{
\item{GEOobject}{GEOobject}

\item{platform}{platform}

\item{TCGAtumor}{tumor name}
}
\value{
return GEO gset
}
\description{
This function retrieves and prepares GEO data
}
\examples{
dataDEGs <- DPA(dataFilt = dataFilt, dataType = "Gene expression")
# to change with get.data.GEO
}

