% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{GSEA}
\alias{GSEA}
\title{GSEA}
\usage{
GSEA(DEGsmatrix, top, plot = FALSE)
}
\arguments{
\item{DEGsmatrix}{DEGsmatrix output from DEA such as dataDEGs}

\item{top}{is the number of top BP to plot}

\item{plot}{if TRUE return a GSEA's plot}
}
\value{
return GSEA result
}
\description{
This function carries out the GSEA enrichment analysis.
}
\examples{
dataDEGs <- DEGsmatrix
# dataFEA <- GSEA(DEGsmatrix = dataDEGs)
}
