% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonlightR.R
\docType{data}
\name{knownDriverGenes}
\alias{knownDriverGenes}
\title{Information on known cancer driver gene from COSMIC}
\format{
A 101x3 matrix
}
\usage{
data(knownDriverGenes)
}
\value{
a 101x3 matrix
}
\description{
A data set containing the following data:
}
\details{
\itemize{
\item TSG known tumor suppressor genes
 \item OCG known oncogenes
}
}
\keyword{datasets}
