% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonlightR.R
\docType{data}
\name{listMoonlight}
\alias{listMoonlight}
\title{Output list from Moonlight}
\format{
A Large list with 5 elements
}
\usage{
data(listMoonlight)
}
\value{
output from moonlight pipeline
}
\description{
A list containing the following data:
}
\details{
\itemize{
	\item listMoonlight output from moonlight's pipeline containing dataDEGs, dataURA, listCandidates
}
}
\keyword{datasets}
