% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPA.R
\name{LPA}
\alias{LPA}
\title{LPA}
\usage{
LPA(dataDEGs, BP, BPlist)
}
\arguments{
\item{dataDEGs}{is output from DEA}

\item{BP}{is biological process}

\item{BPlist}{is list of genes annotated in BP}
}
\value{
table with number of pubmed that affects, increase or decrase genes annotated in BP
}
\description{
This function carries out the literature phenotype analysis (LPA)
}
\examples{
\dontrun{
data(DEGsmatrix)
BPselected <- c("apoptosis")
BPannotations <- DiseaseList[[match(BPselected, names(DiseaseList))]]$ID
dataLPA <- LPA(dataDEGs = DEGsmatrix[1:50,],
                 BP =  BPselected,
                 BPlist = BPannotations)
}
}

