% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEA.R
\name{PEA}
\alias{PEA}
\title{PEA}
\usage{
PEA(listMoonlight)
}
\arguments{
\item{listMoonlight}{output Moonlight function}
}
\value{
no return value, PEA result is plotted
}
\description{
This function carries out the pathway enrichment analysis (FEA)
}

