% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCircos.R
\name{plotCircos}
\alias{plotCircos}
\title{plotCircos}
\usage{
plotCircos(listMoonlight, listMutation = NULL, additionalFilename = NULL,
  intensityColOCG = 0.5, intensityColTSG = 0.5, intensityColDual = 0.5,
  fontSize = 1)
}
\arguments{
\item{listMoonlight}{output Moonlight function}

\item{listMutation}{listMutation}

\item{additionalFilename}{additionalFilename}

\item{intensityColOCG}{intensityColOCG}

\item{intensityColTSG}{intensityColTSG}

\item{intensityColDual}{intensityColDual}

\item{fontSize}{fontSize}
}
\value{
no return value, plot is saved
}
\description{
This function visualize the plotCircos
}
\examples{
plotCircos(listMoonlight = listMoonlight, additionalFilename = "_ncancer5")
}
