% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFEA.R
\name{plotFEA}
\alias{plotFEA}
\title{plotFEA}
\usage{
plotFEA(dataFEA, topBP = 10, plotNAME = "test", height, width,
  offsetValue = 5, angle = 90, xleg = 35, yleg = 5, minY = -5,
  maxY = 10)
}
\arguments{
\item{dataFEA}{dataFEA}

\item{topBP}{topBP}

\item{plotNAME}{plotNAME}

\item{height}{Figure height}

\item{width}{Figure width}

\item{offsetValue}{offsetValue}

\item{angle}{angle}

\item{xleg}{xleg}

\item{yleg}{yleg}

\item{minY}{minY}

\item{maxY}{maxY}
}
\value{
no return value, FEA result is plotted
}
\description{
This function visualize the functional enrichment analysis (FEA)'s barplot
}
\examples{
dataFEA <- FEA(DEGsmatrix = DEGsmatrix)
plotFEA(dataFEA = dataFEA,plotNAME = "FEAplot",height = 20,width = 10)
}

