% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRA.R
\name{PRA}
\alias{PRA}
\title{Pattern Recognition Analysis (PRA)}
\usage{
PRA(dataURA, BPname, thres.role = 0, seed = 12345)
}
\arguments{
\item{dataURA}{output URA function}

\item{BPname}{BPname}

\item{thres.role}{thres.role}

\item{seed}{seed value}
}
\value{
returns list of TSGs and OCGs when biological processes are provided, otherwise a randomForest based classifier that can be used on new data
}
\description{
This function carries out the pattern recognition analysis
}
\examples{
dataDEGs <- DEGsmatrix
dataGRN <- GRN(TFs = rownames(dataDEGs)[1:100], 
DEGsmatrix = dataDEGs,
DiffGenes = TRUE,
normCounts = dataFilt)
dataURA <-URA(dataGRN = dataGRN,
DEGsmatrix = dataDEGs, 
BPname = c("apoptosis",
"proliferation of cells"))
}

