% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roma.R
\name{Roma}
\alias{Roma}
\title{Roma}
\usage{
Roma(exp_data, signature, path)
}
\arguments{
\item{exp_data}{expression data where the activity of the signatures needs to be evaluated}

\item{signature}{in gmt format}

\item{path}{absolute path to MoonlightR folder}
}
\value{
matrix of roma scores
}
\description{
This function carries out the ROMA analysis
}
\examples{
dataDEGs <- DPA(dataFilt = dataFilt, dataType = "Gene expression")
# to change with ROMA
}

