% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunmoon_ill.r
\name{sunmoonIlluminance}
\alias{sunmoonIlluminance}
\title{Get sun and moon illuminance}
\usage{
sunmoonIlluminance(lon, lat, date, TZ, sky)
}
\arguments{
\item{lon}{Longitude in decimal degrees, numeric}

\item{lat}{Latitude in decimal degrees, numeric}

\item{date}{Date (including hour of the day) in a format that is recognized by function ymd_hms of the lubridate package, character string or date
time in Universal Time (UTC) should be preferred (and parameter TZ set to 0)}

\item{TZ}{Time Zone of date parameter (0 => UTC ; 1 => standard zone time ; 2=> local mean time), integer}

\item{sky}{Sky condition (1 => sun/moon visible, cloud cover < 70\% ; 2 => sun/moon obscured by thin clouds ; 3 => sun/moon obscured by average clouds ; (...) 10=> sun/moon obscured by dark stratus clouds)
illuminance is divided by the number provided.}
}
\value{
a list with the following elements:
\itemize{
  \item $sun_illuminance : sun illuminance (lux)
  \item $sun_azimuth : sun azimuth (deg)
  \item $sun_alt : sun altitude (deg)
  \item $moon_illuminance : moon illuminance (lux)
  \item $moon_azimuth : moon azimuth (deg)
  \item $moon_alt : moon altitude (deg)
  \item $moon_pct_ill : proportion of moon illuminated (\%)
  \item $total_illuminance : total illuminance (lux)
}
}
\description{
Get sun and moon illuminance
}
\examples{
# Test cases for program certification (see https://ui.adsabs.harvard.edu/abs/1987USNOC.171.....J p.15)
sunmoonIlluminance(-135.8,-23.4,"19861218210000",0,1)
sunmoonIlluminance(39.5,21.3,"19880813183100",2,1)

# various date format
sunmoonIlluminance(-135.8,-23.4,"1986-12-18 21:00:00",0,1)
sunmoonIlluminance(-135.8,-23.4,"1986-12-18-21-00-00",0,1)
sunmoonIlluminance(-135.8,-23.4,"1986/12/18 21:0000",0,1)



}
