% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonlightR.R
\docType{data}
\name{tabGrowBlock}
\alias{tabGrowBlock}
\title{Information growing/blocking characteristics for 101 selected biological processes}
\format{
A 101x3 matrix
}
\usage{
data(tabGrowBlock)
}
\value{
a 101x3 matrix
}
\description{
A data set containing the following data:
}
\details{
\itemize{
\item tabGrowBlock matrix that defines if a process is growing or blocking cancer development, for each 101 biological processing 
}
}
\keyword{datasets}
