% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URA.R
\name{URA}
\alias{URA}
\title{URA Upstream Regulator Analysis}
\usage{
URA(dataGRN, DEGsmatrix, BPname, nCores = 1)
}
\arguments{
\item{dataGRN}{output GNR function}

\item{DEGsmatrix}{output DPA function}

\item{BPname}{biological processes}

\item{nCores}{number of cores to use}
}
\value{
an adjacent matrix
}
\description{
This function carries out the upstream regulator analysis
}
\examples{
dataDEGs <- DEGsmatrix
dataGRN <- GRN(TFs = rownames(dataDEGs)[1:100], 
DEGsmatrix = dataDEGs,
DiffGenes = TRUE,
normCounts = dataFilt)
dataURA <-URA(dataGRN = dataGRN,
DEGsmatrix = dataDEGs, 
BPname = c("apoptosis",
"proliferation of cells"))
}
