% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataGEO.R
\name{getDataGEO}
\alias{getDataGEO}
\title{getDataGEO}
\usage{
getDataGEO(GEOobject = "GSE39004", platform = "GPL6244", TCGAtumor = NULL)
}
\arguments{
\item{GEOobject}{GEOobject}

\item{platform}{platform}

\item{TCGAtumor}{tumor name}
}
\value{
return GEO gset
}
\description{
This function retrieves and prepares GEO data
}
\examples{
\dontrun{
dataGEO <-  getDataGEO(GEOobject = "GSE20347",platform = "GPL571")
}
}
