% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCircos.R
\name{plotCircos}
\alias{plotCircos}
\title{plotCircos}
\usage{
plotCircos(listMoonlight, additionalFilename = NULL)
}
\arguments{
\item{listMoonlight}{output Moonlight function}

\item{additionalFilename}{additionalFilename}
}
\value{
no return value, plot is saved
}
\description{
This function visualize the plotCircos
}
\examples{
plotCircos(listMoonlight = listMoonlight, additionalFilename = "_ncancer5")
}

