% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFEA.R
\name{plotFEA}
\alias{plotFEA}
\title{plotFEA}
\usage{
plotFEA(dataFEA, topBP = 10, additionalFilename = NULL, height, width,
  offsetValue = 5, angle = 90, xleg = 35, yleg = 5, titleMain,
  minY = -5, maxY = 10, mycols = c("#8DD3C7", "#FFFFB3", "#BEBADA"))
}
\arguments{
\item{dataFEA}{dataFEA}

\item{topBP}{topBP}

\item{additionalFilename}{additionalFilename}

\item{height}{Figure height}

\item{width}{Figure width}

\item{offsetValue}{offsetValue}

\item{angle}{angle}

\item{xleg}{xleg}

\item{yleg}{yleg}

\item{titleMain}{title of the plot}

\item{minY}{minY}

\item{maxY}{maxY}

\item{mycols}{colors to use for the plot}
}
\value{
no return value, FEA result is plotted
}
\description{
This function visualize the functional enrichment analysis (FEA)'s barplot
}
\examples{
dataFEA <- FEA(DEGsmatrix = DEGsmatrix)
plotFEA(dataFEA = dataFEA, additionalFilename = "_example",height = 20,width = 10)
}
