% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotURA.R
\name{plotURA}
\alias{plotURA}
\title{plotURA: Upstream regulatory analysis plot}
\usage{
plotURA(dataURA, plotNAME = "URAplot")
}
\arguments{
\item{dataURA}{output URA function}

\item{plotNAME}{figure name}
}
\value{
heatmap
}
\description{
This function visualizes the URA in a heatmap
}
\examples{
dataDEGs <- DEGsmatrix
dataGRN <- GRN(TFs = rownames(dataDEGs)[1:100], 
DEGsmatrix = dataDEGs,
DiffGenes = TRUE,
normCounts = dataFilt)
dataURA <-URA(dataGRN = dataGRN,
DEGsmatrix = dataDEGs, 
BPname = c("apoptosis",
"proliferation of cells"))
dataDual <-PRA(dataURA = dataURA, 
BPname = c("apoptosis",
"proliferation of cells"),
thres.role = 1)
plotURA(dataURA = dataURA[c(names(dataDual$TSG), names(dataDual$OCG)),],plotNAME = "URAplot")
}

